general commit

+++ event pop-up/ux_definitions.py
@@ -3,6 +3,7 @@ import pickle
 import hmac
 import hashlib
 import datetime
+import os
 
 
 def b_ok(event=None):
@@ -26,22 +27,32 @@ def b_ok(event=None):
     for i in m:
         responses += [i.get('1.0', 'end-1c')]
 
-    with open('Storage.dat', 'rb+') as f:
-        with open('DATA.BIN', 'rb') as g:
-            ke = g.read()
+    if os.path.exists('Storage.dat') and os.path.getsize('Storage.dat'):
+        with open('Storage.dat', 'rb+') as f:
+            with open('DATA.BIN', 'rb') as g:
+                ke = g.read()
 
-        i = f.read()
-        if hmac.new(b'shared-key', i, hashlib.sha256).digest() == ke:
-            di = eval(pickle.loads(i))
-            di[str(datetime.datetime.now().strftime('%d') + datetime.datetime.now().strftime('%m') + datetime.datetime.now().strftime('%Y'))] += [responses]
+            i = f.read()
+            if hmac.new(b'shared-key', i, hashlib.sha256).digest() == ke:
+                di = eval(pickle.loads(i))
+                di[str(datetime.datetime.now().strftime('%d') + datetime.datetime.now().strftime('%m') + datetime.datetime.now().strftime('%Y'))] += [responses]
 
-            x = pickle.dumps(str(di).encode())
+                x = pickle.dumps(str(di).encode())
 
-            with open('DATA.BIN', 'wb') as g:
-                g.write(hmac.new(b'shared-key', x, hashlib.sha256).digest())
+                with open('DATA.BIN', 'wb') as g:
+                    g.write(hmac.new(b'shared-key', x, hashlib.sha256).digest())
 
-            f.seek(0)
+                f.seek(0)
+                f.write(x)
+    else:
+        x = pickle.dumps(str({str(datetime.datetime.now().strftime('%d') + datetime.datetime.now().strftime('%m') + datetime.datetime.now().strftime('%Y')): [responses]}).encode())
+
+        with open('Storage.dat', 'wb') as f:
             f.write(x)
+
+        with open('DATA.BIN', 'wb') as f:
+            f.write(hmac.new(b'shared-key', x, hashlib.sha256).digest())
+
     root.destroy()
